Add task directories
The easy directory includes a patch that adds helpful NOTICE messages
for the user and relatively more detailed instructions in its
README. This directory is recommended for people that are rusty with
Postgres internals or for those with no Postgres internals knowledge
at all.

The hard directory just provides a README that is more open-ended on
how the user can approach the tasks in any way they want. This
directory is recommended for people that are confident in their
Postgres internals knowledge and/or C skills.

Also modified the pg_regress test to match with the README file
example for easier confirmation testing.

Whichever directory chosen though, the main thing is to observe how
the user does their implementation (e.g. code navigation, testing,
communication, etc.). There's no expectation to have the tasks
completed at the end of the session.