Add loadable modules as tests of the amapi
If the index access method API meets the needs of future access
methods, then surely it must at least meet the needs of present
access methods.  Copying and renaming hash and btree as "xash" and
"xtree", respectively, those access methods should perform exactly
the same as the in-core hash and btree methods.  Alas, they do not.
Running the standard in-core regression and isolation tests against
these add-on index access methods fail in some notable ways.

Rather than suffer the maintenance headaches of copying and slightly
modifying the regression and isolation test suites, add a script to
copy such files and insert "USING xtree" and "USING xash" clauses in
the appropriate places.

The regression tests fail after applying this patch.  That's the
point of the patch, not an oversight.  To not be a burden, these
modules are skipped during `make check-world`, and the directories
are entirely omitted from the meson build system.