Add "treeb" deep copy of btree
This is committed separately from the "xash" and "xtree" of the
previous commit, so that "treeb" is available for reviewing core
changes while separable for the final commit.  This commit and
others that follow which change "treeb" can be excised from the
final patch set if "treeb" is unwanted.

The name "treeb" is a rotation of "btree".  The idea is that the
StrategyNumbers for LT, LE, EQ, GE, and GT will also be rotated,
though the commit to do so is delayed until sufficient
infrastructural changes have been made to support such a change.
(Currently, too much of the system assumes that treeb must have the
same strategy numbers as btree.)

Although this patch series ends with a commit which performs the
strategy rotation, assuming "treeb" is committed, future work can be
done to test other aspects of the amapi by making additional
changes.  This is something that cannot readily be done using the
shallow copies.