Convert strategies to and from row compare types
For each Index AM, provide a mapping between operator strategies and
the system-wide generic concept of a row comparison type.  For
example, for btree, BTLessStrategyNumber maps to and from
ROWCOMPARE_LT.  Numerous places in the planner and executor think
directly in terms of btree strategy numbers (and a few in terms of
hash strategy numbers.)  These should be converted over subsequent
commits to think in terms of RowCompareType instead.