Fix direct cast from strategy to rctype
Prior to this patchset, the code had a single cast from an index
strategy to row compare type.  That worked, because only btree was
allowed, and the row compare type values were designed to be equal
to btree strategy numbers.  This no longer works, as we want to
allow other index access methods.